+++ packages/devtools_app/lib/src/shared/release_notes/release_notes.dart
@@ -11,15 +11,15 @@ import 'package:flutter/material.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:http/http.dart' as http;
 
-import '../../devtools.dart' as devtools;
-import '../config_specific/launch_url/launch_url.dart';
-import '../config_specific/logger/logger.dart' as logger;
-import '../config_specific/server/server.dart' as server;
-import '../primitives/auto_dispose_mixin.dart';
-import 'common_widgets.dart';
-import 'theme.dart';
-import 'utils.dart';
-import 'version.dart';
+import '../../../devtools.dart' as devtools;
+import '../../config_specific/launch_url/launch_url.dart';
+import '../../config_specific/logger/logger.dart' as logger;
+import '../../config_specific/server/server.dart' as server;
+import '../../primitives/auto_dispose_mixin.dart';
+import '../common_widgets.dart';
+import '../theme.dart';
+import '../utils.dart';
+import '../version.dart';
 
 const debugTestReleaseNotes = true;
