+++ tool/README.md
@@ -95,14 +95,6 @@ git rebase-update
 third_party/devtools/update.sh 8881a7caa9067471008a8e00750b161f53cdb843
 ```
 
-### Publish package:devtools_shared on pub
-
-`package:devtools_shared` is the only DevTools package that is published on pub.
-From the devtools/packages/devtools_shared directory, run:
-```shell
-pub publish
-```
-
 ### Update the DevTools hash in the Dart SDK
 
 Update the `devtools_rev` entry in the Dart SDK 
@@ -119,3 +111,18 @@ gclient sync -D
 ./tools/build.py -mrelease -ax64 create_sdk
 out/ReleaseX64/dart-sdk/bin/dart devtools  # On OSX replace 'out' with 'xcodebuild'
 ```
+
+### Publish package:devtools_shared on pub
+
+`package:devtools_shared` is the only DevTools package that is published on pub.
+From the devtools/packages/devtools_shared directory, run:
+```shell
+pub publish
+```
+
+### Write release notes for the release
+Release notes should contain details about the user-facing changes included in the release.
+These notes are shown directly in DevTools when a user opens a new version of DevTools. Please
+see the release notes
+[README.md](https://github.com/flutter/devtools/blob/master/packages/devtools_app/lib/src/shared/release_notes/README.md)
+for details on where to add release notes and how to test them.
